requirements.txt
requests==2.31.0 
lxml==4.9.3 
psycopg2-binary==2.9.9 
python-dotenv==1.0.0 
playwright==1.47.0